requirements.txt
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
Flask==2.1.2
functions-framework==3.0.0
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-bigquery==3.1.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-pubsub==2.13.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
oauth2client==4.1.3
packaging==21.3
pandas==1.4.2
proto-plus==1.20.3
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
watchdog==1.0.2
Werkzeug==2.1.2
zipp==3.8.0
